Moved configure output options test to configure module from core.
